+++ derive_builder_test/src/lib.rs
@@ -1,3 +1,5 @@
+#![feature(try_from)]
+
 #[macro_use]
 extern crate derive_builder;