+++ derive_builder/CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+- Implement `TryFrom` from builder to target type from generated builder 
+  when `try_from` crate feature is enabled.
+
 ## [0.4.4] - 2017-04-12
 
 ### Added